refactor: AlbumInfo 부수 컴포넌트 파일로 정리
