+++ client/src/entities/songDetail/index.ts
@@ -1,0 +1,2 @@
+export { LyricsPanel } from './ui/LyricsPanel';
+export { PlaylistPanel } from './ui/PlaylistPanel';
