+++ client/src/entities/songDetail/ui/PlaylistPanel.tsx
@@ -1,0 +1,3 @@
+export function PlaylistPanel() {
+  return <p>플레이리스트</p>;
+}
