+++ client/src/features/songDetail/index.ts
@@ -1,0 +1,1 @@
+export { SongDetail } from './ui/SongDetail';
