+++ client/src/features/songDetail/model/types.ts
@@ -1,0 +1,4 @@
+export const CATEGORIES = {
+  LYRICS: 'lyrics',
+  PLAYLIST: 'playlist',
+} as const;
